requirements.txt
asgiref==3.5.1
certifi==2021.10.8
distlib==0.3.4
Django==3.2.9
django-cors-headers==3.11.0
djangorestframework==3.13.1
filelock==3.6.0
gunicorn==20.1.0
Pillow==9.1.0
pipenv==2022.5.2
platformdirs==2.5.2
psycopg2==2.9.3
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
virtualenv==20.14.1
virtualenv-clone==0.5.7
